requirements.txt
click==8.0.1
Envelopes=0.4
pyinstaller==4.5.1
python-dotenv==0.19.0
